create draft release on tag, publish windows and pecl builds (#109)
we already build windows binaries, so let's make them available to users. On each github action, the binaries are published as build artifacts.
When a tag is created, we now create a draft release from the workflow (instead of manually creating the release via the GUI), which means we can also upload the binaries as release artifacts.
Give PECL package the same treatment, so that maintainers don't have to build it manually and upload.
After this is merged, the release process for the extension will change slightly: it relies on a tag being created (but not a release). You can't do that from the GUI, so there's a change coming to dev-tools which will create the tag, and from there the workflow will take care of creating the release in draft form. The final check is for a maintainer to eyeball the release, make any modifications to the text, and publish it.