observe methods from attribute (#146)
Add the capability to observe/hook methods and functions via WithSpan and SpanAttribute
attributes. This capability is disabled by default (it has some runtime overhead), and can be enabled via php.ini

WithSpan is a signal that a function or method should be auto-instrumented. SpanAttribute enables arguments to be passed through to pre hook as attributes to be added to a span.
The attributes are provided by the API, as OpenTelemetry\API\Instumentation\WithSpan and OpenTelemetry\API\Instrumentation\SpanAttribute
Two extra parameters are passed to pre hooks:
1. span name + span kind (from WithSpan's arguments, if provided)
2. attributes (from WithSpan's 3rd argument, and from SpanAttribute)

WithSpan and SpanAttribute can be applied to a methods, functions, or interface methods.

The default pre/post hook callbacks for WithSpan are provided by the API, as OpenTelemetry\API\Instrumentation\WithSpanHandler. They can be changed via php.ini

There is a restriction that WithSpan hooks can only be added to functions/methods that
are not already hooked by something else (because the hooks are added at runtime: when
a method is executed and has no hooks, we then check for attribute and add hooks.