fix observing namespaced functions (#200)
* fix observing namespaced functions
if a user wants to observe a namespaced function, this allows adding the observer
with or without a leading slash. Previously, the leading slash caused the lookup to fail,
which is not clear to users. Added a test to show that it works with both forms now.

* avoid slash direction for windows

* strip slash in add_function_observer
per review feedback

* verify that classes can be hooked with/out leading slash